Modified stimulus amplitude so that they are referred in terms of attenuation relative to Short.MaxValue.
Playback setting is always set at maximum value. All 'volume' changes should be done at the signal amplitude level and relative to the Short.MaxValue.

+++ trunk/src/org/audiopulse/io/RecordThreadRunnable.java
@@ -74,7 +74,7 @@ public class RecordThreadRunnable implements Runnable
 	{
 		Log.v(TAG,"constructing record thread");
 		mainThreadHandler = h;
-		Buffer_Size=(int) (playTime*(double)sampleRate);
+		Buffer_Size=(int) (playTime*sampleRate);
 		samples = new short[Buffer_Size];
 		initRecord();
 		IN_REC_MODE=0;
